final/requirements.txt
flask
google-cloud-datastore
gunicorn
google-cloud-storage
google-cloud-vision
google-cloud-translate
google-cloud-texttospeech